Correctly materialize and connect fifos with non-default reset name.
The reset port of a FIFO instantiation is always called
`FifoInstantiation::kResetPortName`. The codegen backend, however,
allows for an arbitrary reset signal name to be specified for blocks.

This patch renames the reset port of a FIFO instantiation during
materialization to the name specified in the codegen options, so that
the block behaves like a 'normal' block created during codegen and does
not require any special handling.